requirements.txt
azure-functions==1.24.0
azure-data-tables==12.7.0
azure-identity==1.25.1
